Avoid waiting perpetually for DNS resolution
Fixes #6995
